refactor: rename preview client to core
The client maps to the core API service, not specifically to previews.
