+++ src/commands/preview-add.ts
@@ -1,7 +1,7 @@
 import { parseArgs } from "node:util";
 
 import { getHost, getToken } from "../auth";
-import { addPreview } from "../clients/preview";
+import { addPreview } from "../clients/core";
 import { UnknownRequestError } from "../lib/request";
 import { safeGetRepositoryName } from "../project";
